Create two separate form components apart from page logic
